feat: add useGetPrice and useSelectReducer hooks
